requirements.txt
aiogram
SQLAlchemy
psycopg2-binary
python-dotenv
openai
